[CELEBORN-169][REFACTOR] Extract ObservedDevice out from LocalDeviceMonitor
